applied patch 2878824: The p2os driver doesn't report any position changes

+++ server/drivers/mixed/p2os/robot_params.h
@@ -103,7 +103,7 @@ typedef struct
   int CollisionRange; //
   double DiffConvFactor; //
   double DiscardThreshold; //
-  int DistConvFactor; //
+  double DistConvFactor; //
   double DistanceWt; //
   int DrivingRotAccel; //
   int DrivingRotDecel; //
